Remove placement columns from application reports
Placement columns have removed from the applications and referrals (aka combined application and placement) reports.  These are now available from the new placement matching outcome reports.

For context:

“The raw metrics reports should only contain information that is relevant to the period in which they are being downloaded for. If they contain too much information then users may try to create metrics based on datasets that are incomplete.

Matching and Booking data should be removed from reports until the user is ready to report on those”

The reports will only ever return a single row per application

The expected end date is now populated from the ‘initial’ placement request that is created once assessment is complete. This is because until this point the duration being used for the stay is not known/confirmed