Tidy up tests
This makes the tests easier to grok. I’ve also added a `perPage`
argument so we can test pagination without having to create loads
of objects in the database.

It would be good if we could tweak the test config so we can optionally
keep the database around for the lifecycle of each inner class (so we
don’t have to recreate the objects each time), but that’s a rabbit hole
for another day.