Don’t link adhoc placements to match requests
Before this commit an adhoc placement would be linked to the latest ‘non placed’ match request on an application with the same CRN. This could result in placements being linked to match requests with completely different dates.

As the link to match requests is not neccessary, this functionality has been removed.