Adhoc bookings are exempt from cascade withdrawals
Several legacy adhoc bookings are linked to potentially unrelated placement requests. This behaviour was removed in commit 2d96a4d37567fde3f91a7a9172b459a91fb30625.  To avoid such adhoc placements being unintentionally withdrawn when (unrelated) ancestor elements are withdrawn, we mark them as exempt from cascade.

Exempting such makes sense in the general context because when creating an adhoc booking the user doesn’t explicitly link it to an application, it is automatically assigned to an application with the required CRN (if such an applciation exists). Therefore, it’s unexpected for an adhoc to be withdrawn automatically if an application withdraws the application.