Populate CAS1 domain_event user & add to timline
Populate domain_event.triggered_by_user_id for CAS 1 only.

Add user to CAS1 Domain Event Timeline Entries

This commit also changes the DomainEventSummary query to use JPA QL instead of native SQL, allowing us to leverage hibernate to fully populate the UserEntity returned. Because of this, we can now use OffsetDateTime in DomainEventSummary instead of java.sql.Timestamp

The ‘service’ set in UserTransformer is now correctly set to “CAS1”, matching the discriminator defined for the User Type in the Open API Spec. This was required to allow tests to correctly unmarshal user json directly using the Jackson ObjectMapper (this was specifically an issue in AppealsTest.`Create new appeal adds an event to the application timeline`)