Allow many ‘initial’ match requests to be returned
Whilst there should only ever be 0 or 1 placement requests returned when requesting getPlacementRequestForInitialApplicationDates, due to data inconsistencies there are a small number of applications that will return multiple placement requests.

We know that only one of these placement requests relate to the initial application dates, and the rest relate to placement_applications, but we have no programmatic way of 'fixing' this link for those placement requests.