add Offender service method to get person's name
We would like to move the person in prison's name to a top level field on an Application Summary, as part of a larger refactoring work to avoid making continual calls to integration APIs to retrieve this data.

Currently on CAS2 we are making a call to Delius to get a `personInfoResult` which returns lots of fields relating to a Person when all we need is a name. This is a temporary method to extract the name from the Result which will be used in the SubmissionsController.