return `personName` as top level field on SubmittedApplicationSummary
We would like to move the person in prison's name
to a top level field on a Submitted Application
Summary, as part of a larger refactoring work to
avoid making continual calls to integration APIs
to retrieve this data.

Currently on CAS2 we are making a call to Delius
to get a `personInfoResult` which returns lots of
fields relating to a Person when all we need is a
name.

We do not want to break the frontend contract, so
here we keep the `person` field and add in
`personName`. In the near future when the frontend
has been updated to read `personName` we will be
able to remove `person` and refactor the
`getInfoForPersonOrThrowInternalServerError` to
make only one call to
`offenderDetailsDataSource.getOffenderDetailSummary`
rather than also calling
`prisonsApiClient.getInmateDetailsWithWait` to get
the full person details.