Use ‘adhoc’ boolean when building withdrawal tree
Previously we used the abence of a placement_request associated with a booking to determine if it’s adhoc. For some legacy applications we have linked placement_requests to adhoc placements, so this is not a reliable indicator.

Therefore, we instead use the ‘adhoc’ property to find adhoc bookings. If the ‘adhoc’ property is null it’s unknown if the placement is adhoc, so we assume it is (as this is the least destructive way to handle a booking wrt. withdrawal as it limits cascading)