Check to ensure withdrawal is limited to same app
We previously had a bug in the code where withdrawing an application would inadvertently cascade to adhoc bookings from other applications.

Whilst this bug has been fixed and specific regression tests added for the specific scenario, we want to add additional guard rails in the cascade withdrawal logic to ensure is not possible for cascades of any entity types. This is especially important as withdrawal cascades are somewhat opaque on the UI as the user it not explicitly told which entities will be withdrawn on cascade.

This commit adds a check in the withdrawal logic to ensure that all entities being withdrawn belong to the same application as the root entity being withdrawn.

Note that this implementation isn’t perfect as the check only occurs after the root entity has already been withdrawn (ideally this would be blocked). This is acceptable because this scenario should never occur, and the guardrails are only in place to avoid very damaging side affects, which _would_ be blocked