remove `Person` on Cas2SubmittedApplicationSummary
as we now have the `personName`, `crn`,
`nomsNumber` fields as top level fields, we can
simplify how we are getting the person's name from
the `offenderDetailsDataSource`.

I'm also slightly changing the error behaviour -
previously we called the function
`getInfoForPersonOrThrowInternalServerError` which
would throw an error if there person was 'Not
Found'. For our purposes here in CAS2, I don't
think we need to throw an error at this point - if
the person is not found on Nomis/Delius then we
still want to go ahead and show the application,
but with placeholder copy. It could be that the
application still needs to be processed or viewed
even if the applicant is no longer in the system.