Fix Placement App Links Job Exclusion Fix
Previously the migration job would error if an application was returned from the initial query where there was at least one placement application with a decisionMadeAt date set. This was a safety check to ensure we were not modifying placement applications that were created after the issue this backfill was introduced for was resolved (this is because decisionMadeAt was only populated in the code after we’d solved the issue that led to this backfill being required)

This check didn’t consider that old applications can still have new placement applications added to them, so several applications were being incorrectly excluded. The logic now only errors if ALL placement applications have decision made at set.