Set noticeType at the point of creation
This supports both the setting of a specific noticeType field, as well
fetching it from the presnece / absence of `isEmergencyApplication` and
the result of `isShortNoticeApplication`. Once we update the frontend,
this complexity can be removed and we can rely more on the frontend to
tell us an application’s notice type.