check all CAS2 reports for role-based controls
here we loop through a list of the reports that
are possible to download from CAS2 and check that
users without correct roles cannot access them.

Because we do not have unit testing around our
oauth configuration in
`OAuth2ResourceServerSecurityConfiguration` then
I'm not sure how else we could have confidence
that every end point is covered.