Generate an ID for documents without an ID
Before we couldn’t rely on the presence of an ID for a document, as the
API didn’t always return them. This adds a virtual `_id` field and an
`id` getter that returns the virtual field if it exists, otherwise it
returns a UUID generated from a byteArray of the document’s title
and the created at date. Generally we can assume that documents
will have unique titles, but adding the createdAt date gives us that
little bit more confidence.

I’ve also added some unit tests for `findDocument` to ensure that
fetching documents by their ID still works with autogenerated IDs.

We also need to add a `JsonProperty` annotation to the virtual field,
so Jackson knows to convert any JSON `id` keys to the virtual `_id`
key