Test that if prison API fails we return validation error
I think it makes sense to handle this as
validation error, because it means required data
we need to validate application is missing.

Here we also add our own inmateDetail to the
`given and Offender` function, as the standard
inmate detail does not include the prison code we
now require.