Add job to fix broken ‘further info’ assessments
There is a known issue that affects assessments where a user has requested further information, and goes back to edit the respons, selecting ‘no’ beacause they don’t want to trigger a re-request. This inconsistent state leads to an infinite loop when the UI is attempting to validate the answers.

This commit introduces a seed job the updates the ‘sufficient information’ response to ‘yes’ to break the infinite loop. It can be used to ‘fix’ assessments in this inconsistent state