Add `latestStatusUpdate` to CAS2ApplicationSummary
This will be used to display the status of an Application's in the
Referrer dashboard. It's possibly a little confusing that we have this
field in addition to the `status` field, but we're planning to refactor
status updates onto an `Assessment` in the near future so we've decided
to put them here for now.