WIP Return latest status update on ApplicationSummary
This will be `null` for unsubmitted applications, or those with no
ApplicationSummary.

This isn't working - I've spent most of the day working on the SQL query
here as I believe we should be joining the Application and StatusUpdate
tables, as my previous attempts before joining would always return a
`null` status update.

I'm not exactly sure what the problem is here, but the SQL seems to be
working as expected when I've tried it in the IntelliJ console,
returning Applications with their latest StatusUpdate.