On upsert set user apArea & teamcodes
When a user is created or updated, if the request originated from a CAS1 UI we should also populate the user’s ap area.

This commit also ensures team codes are always captured on creation and update.

Whilst backfill jobs have shown that AP Area can be determined for all users currently registered in the system, in the future we do not want the inability to populate AP Area to block a user’s access to CAS2 or CAS3