Change how user’s area is resolved
Previous to this commit we would retrieve the user’s ap area from their corresponding probation area (e.g. user.probationArea.apArea). This mapping is not correct for all users, as the team code should also be considered.

This commit introduces a new ap_area_id column on the users table which is populated whenever the user is created/updated via a request originating from the CAS1 UI. This ap area is returned to the UI in the User Model and used for user searches.