Change how user’s area is resolved
Previous to this commit we would retrieve the user’s ap area from their corresponding probation area (e.g. user.probationArea.apArea). This mapping is not correct for all users, as the team code should also be considered.

This commit uses user.ap_area_id column instead to determine which AP Area should be used for a user.

This commit also updates the test user’s SQL used in local/dev/test to pre-populate the ap area code to the value corresponding to that previously retrieved from the probation region for the given user. This resolves an issue in local environments where the ap area cannot be correctly retrieved due to a mismatch in data used for local deployments of community api vs the prod deployment. This issue will be correctly resolved via changes for APS-579