use OffsetDateTime for cas2 submitted report
We were previously using the `Instant` created on
the event data, which did not take into account
timezones. Using the `occurred_at` means we are
using an OffsetDatetime/ `TIMESTAMP WITH TIMEZONE`
which is the same format used for
`application.created_at` which means it should
reflect the correct local time at which the event
happened. It also means we don't have to dig into
the JSON data to find the the timestamp.

This change is happening in the SQL query itself -
we don't have unit tests for these queries, and
it's difficult to test in integration tests, see
this commit for information
https://github.com/ministryofjustice/hmpps-approved-premises-api/commit/260b1c6ded66c71d8495eca6512d9663be0dfe9e

But I think we have enough confidence by checking
that the `createdAt` of the report matches the
`occurredAt` of the event entity.

I've also tidied up the tests to use less verbose
means of getting timestamp.