Update ApArea backfill to always populate ap area
Previously the ap area was not populated if the staff details could not be found. Because the user.ap_area_id is required to show the users on the users page, we really need this to be populated with something, even for inactive users.

This change updates the ap area backfill to fall back to probationRegion.apArea where the staff user details cannot be retrieved