Populate ap_area_id for any new user, regardless of the CAS
Previously we only populated users.ap_area_id for CAS1 users, because we didn’t want logins to fail for CAS3 users if the ap area couldn’t be resolved.

We then realised that this caused an issue when viewing the user management page, as this page list users regardless of which CAS they first logged into, meaning some of these users would not have an ap_area_id value defined. This led to errors as all CAS1 users are assumed to have an ap_area_id defined.

This change ensures that users.ap_area_id will always be populated. A more long term solution is required to ensure that CAS3/2 logins do not have to resolve users.ap_area_id as this is unreleated to these systems