Synch user details with Nomis User Roles API every login
We will shortly be basing permissions on a user's
activeAgencyId (prison code). This could change
over time or if a person could have multiple
prisons and change their active one manually, so
we need to check that we have the correct one
every time they login.

This gives us added benefit that we can also check
other data like name and email is still correct.