create LatestStatusUpdate
The UI needs to assign colour/logic to status
updates - the label for a status update could
change over time, because it's the human-readable
string. This commit now passes the UUID for the
status update, set by the
`Cas2ApplicationStatusSeeding` because this will
remain constant. To do this we get the two fields
in the SQL query and then use the Transformer to
transform them to a nest object on the
ApplicationSummary.