Add an `isCompleted` argument to /tasks
This allows us to optionally return completed tasks instead of
incomplete tasks (which is the default).

This makes the SQL a tad more complex. When we have
time we should really think about refactoring to have actual
Task entities which map to the relevant Assessment / Placement
Application / Placement Request, but this is probably OK
for now.