Allow users to see submitted applications within the same prison
Here we introduce a check for users to see
applications if the offender at the time of
submitting the application had the same prison
code associated with it.

Note: As per the Nomis-User-Roles API, the nomis
user's prison code is called `activeCaseloadId`,
and could be null (though we haven't come across
user in production without one).

We are here using the `referringPrisonCode` saved
on our Application - this is currently saved at
the time when the application is submitted, so
this new permission will only apply to submitted
applications. I am also explicitly checking that
the application has been submitted, in case the
logic around setting that referral code changes,
as we do not want this to apply to in progress
applications at the moment.

There may be requirement in the future to update
this `referringPrisonCode` if, for example, the
offender has been moved between prisons after the
application has been submitted. For now though, we
are using the 'snapshot' of the code taken at the
time of submission.