CAS1 App Accepted creates PR Create Domain Event
When an application is assessed and accepted, a placement request is automatically created.

This commit ensures that a ‘Request for Placement’ Domain Event is also raised to ensure this event is reflected in the timeline.

We discussed creating this event when the applicant initially submits the application as this is technically when they’re making a request for a specific date, but as the duration can be modified when being assessed we decided it was best to add to the timeline when assessment is accepted.