Allow users to add notes for apps within same prison
CAS2 users should be able to add notes for
submitted applications where the
`referringPrisonCode` is the same as the user's
`activeCaseloadId`.

Here we introduce a check for this, re-using the
previously created function on the
`UserAccessService.`

There may be requirement in the future to update
this `referringPrisonCode` if, for example, the
offender has been moved between prisons after the
application has been submitted. For now though, we
are using the 'snapshot' of the code taken at the
time of submission.