refactor Note Service
Before adding tests for functionality around prison permissions,
this refactors the order of checks and tests that we make.
1. Return early if application not submitted. I think it makes more sense to return
as soon as possible if application is not submitted, this also makes our unit tests simpler.
2. Create nested classes for possible conditions rather than successful/unsuccessful states.
This organises tests in a more semantic/behaviour driven way, hopefully making the user needs
more readable.