add missing fields to seeded CAS2 applications
This adds missing 'first class' fields to
applications seeded by the `csv` data in
`Cas2ApplicationSeedJob`. I have set individual
applications with their own `referringPrisonCode`
so that we can in future test that users are able
to see submitted applications within their own
prison, and stopped from seeing them for different
prisons.

`BAI` is the code used by our local POM_USER,
which is set in the ap-tools repo. We can set this
as the active caseload for users in the dev/test
environment via the `manage-users` service[1] by
adding the prison to the caseload of the user via
an admin account, and then logging into DPS[2] as
the user and setting it as activecaseload.

The other fields are currently not used for
testing so I think can just be set across all
applications.

[1] https://manage-users-dev.hmpps.service.justice.gov.uk/
[2] https://dps-dev.prison.service.justice.gov.uk/