Use a suppression list for email errors
Previously we raised sentry alerts for a specific error message (Missing personalisation). There are potentially other error messages that we do want raising in Sentry that we’re not aware of.

This commit ‘flips’ the logic used to determine if an email error is raised in sentry, using a supression list to stop unknown and harmless errors being raised in sentry