Allow integration tests to be run per-class
This adds a new `PerClassIntegrationTestBase` abstract class, which
changes the lifecycle of a test class, so we don’t have to setup and
teardown the database for each test, allowing scope to be shared between
tests, so we’re not creating / deleting objects each time.

I’ve changed the IntegrationTestBase to have two unified setup/teardown
functions, which run `@BeforeEach`. These are skipped if the
`isPerClass` tag is present. In the new test base, we have `@BeforeAll`
and `@AfterAll` functions which delegate to the `setupTests` and
`teardownTests` functions.

Additionally, we change the `DBExtension` utils to check if the
tests are running on a Per Class basis. If this is the case, we only run
the `setupDatabase` function at the start of the test run, and skip
it before each individual test.