add custom countQuery to submitted Applications
This query was failing when a pageable object was passed in-
from debugging the hibernate/sql going on, it looks like it was failing
when trying to do the count query, when there was a LEFT JOIN in place.
This adds a custom countQuery that is much simpler than
the select query.

I've also added in a test for using pagination with `isSubmitted`

https://docs.spring.io/spring-data/jpa/docs/current/api/org/springframework/data/jpa/repository/Query.html#countQuery()