Refactor the premises summary tests
There’s a lot of expensive setup in these tests too, so let’s share the
DB context too. I’m not 100% happy with these tests, but I don’t
want to spend too much time on it. Now they’re nicely seperated out
it should be easier to refactor when we next touch this part of the code