DRY up the `DomainEventsController`
This adds a `getDomainEvent` private function which calls the
relevant `domainEventService` method, throwing a `NotFoundProblem`
if there is no domain event. We have to supress the unchecked cast
warning, but I think this is safe as the return type is fixed at runtime
and this is covered in the integration tests.