Update save*Event tests
Now `saveAndEmit` is tested directly, rather than painstakingly testing
each edge case, now we can create a `spyk` version of the service, and
check `saveAndEmit` has been called with the correct arguments.

Arguably this kind of logic could probably be moved to the
`Cas1*DomainEventService` classes, but this is at least an improvement
that we can build on.