Add noms number to domain_event column
To replay a domain event (i.e. adding the necessary information in an SNS message) we need its corresponding noms number, if defined.

Currently the noms number is only stored in the domain event JSON, which is difficult to extract programmatically. This is because we don’t have a common hierarchy for the ‘domain event envelope’ classes as they’re generated from open api specs. This means we’d need ‘type specific’ code to extract the noms number for each domain event type which would be cumbersome.

For that reason, this commit introduces a nullable ‘first class property’ for the noms number on the DomainEventEntity, populating wherever there is a noms number relevant to the domain event for CAS1, 2 and 3