Ensure status is not reset when updating
We’ve noticed that assessments with a status of
`REQUESTED_FURTHER_INFORMATION` were not showing
correctly. After some investigation I discovered that the status
was getting set, but then immediately being set back to
`ASSESSMENT_IN_PROGRESS` with a subsequent update, which
is obviously not what we want.

I’ve updated the AssessmentListener to immediately short circuit
if the application’s status is `REQUESTED_FURTHER_INFORMATION`,
as well as updated the tests to give a more realistic scenario.