give LICENCE_CA create/update permissions
We are adding a new user role to CAS2 -
LICENCE_CA. The current thinking is that this is
the more correct role for our user base than POM,
and so we will replace POM with LICENCE_CA in a
phased approach. Previously we allowed LICENCE_CA
to read applications from their prison, now we are
giving them all the same permissions as POM.
Since our only role-based logic is in
`OAuth2ResourceServerSecurityConfiguration` we
just need to make the change there, and then check
in integration tests that everything is working.