Catch RuntimeExceptions in InformationReceived job
We hit a RuntimeException when running this in preprod, because a user
doesn’t exist. We want to cleanly recover from this and continue.

I’m hoping that this issue won’t come up in prod, but I want to run this
again in preprod and see how many errors we get. If there are only a
handful, I think this will be safe to run in prod. If we get any errors
in prod, we can note down which ones have come up and decide
how to proceed from there.