Increase web client buffer size for prison API
We have an issue in production retreiving alerts from the prison API where the response exceeds our limit of 750000. This commit doubles that limit for prison API only to 1MB.

Note that going forward we should look at other methods to limit the buffer size e.g. stream the response, or see if the API can support paging.