Increase prison api buffer size
This is to unblock a production issue for an application with a large number of alerts.

This is a short term fix and paging/limiting results and/or buffering efficiencies should be considered to resolve this correctly. A new endpoint would be needed on probation-integration context if paging/limiting was requried