Demonstrate that existing roles and qualifications are not removed
Some of the AP staff being seeding will already be presesnt in the service
and some of those will have existing Roles and Qualifications.

We need to be confident that we aren't clearing those Roles and Qualifications.

In fact this is why we are introducing the new `ApStaffUsersSeedJob` rather than
using the `UsersSeedJob` which does clear those roles and qualifications.