Add funcs to detemrine if PR/PA are withdrawable
Currently the code in the isDirectlyWithdrawable function retrieves the complete set of WithdrawableEntities for the whole application, despite only being concerned with WithdrawableEntities for the specific placement request/placement application. This will become a performance issue if/when we start calling external APIs to determine if an entity is withdrawable (as is planned). Before any such change is made we should modify this methods to test against a subset of the tree (i.e. where the root node is placement application/placement request)