Seed CAS1 Application using Auto Script
We previously had some default applications seeded via a ‘local only’ flyway script. Unfortunatley these became stale over time and generally cause issues using the application.

This commit removes those applications and instead bootstraps a single application to the point of submission using an auto script.

Unlike CAS2, we only want the seed job to run locally are it relies on locally managed test data (i.e. CRNs). With improvements this may be usable in other environments. For this reason this commit introduces separate auto script enabled configuration for CAS1 and CAS2

We have elected to use the APIs (or at least, the service function) to create applications, instead of creating them directly in the database. This has several advantages:

* Compared to using SQL, it’s less likely we’ll need to update the auto script when changes are made to CAS1. If the service function is changed the dev will be forced to fix the Auto Script
* The correct domain events will always be created
* It’s easy to read the code and refactor

One downside is that because we rely on APIs to build applications (e.g. retrieve offender details), these need to be running before the API starts. A change to tools has been made to ensure this.