Add new endpoint to provide withdrawal notes
The logic used to determine what is and what isn’t withdrawal is becoming increasingly complex.

To help users understand why something may or may not be withdrawable, a new ‘withdrawablesWithNotes’ endpoint has been added that can be used to provide high level explanations to the user as to why certain elements aren’t withdrawable. Whilst the process still remains reasonably opaque with these high level notes, they at least provides a simple to implement method to explain some of the complexity to the users.