Introduce WithdrawableTree
We have various functions directly returnin an individual WithdrawableTreeNode representing the root node of a tree. This commit introduces a WithdrawableTree that wraps this root node to better represent what is is. Furthermore, it provides a suitable container for producing notes regarding the tree, allowing us to subsequently show notes when logging out the WithdrawableTree, providing a useful debugging tool.

For example:

Application(c2b), withdrawable:Y, mayDirectlyWithdraw:Y, BLOCKED
---> PlacementRequest(20c), withdrawable:Y, mayDirectlyWithdraw:Y, BLOCKED
------> Booking(dc3), withdrawable:N, mayDirectlyWithdraw:Y, BLOCKING - ArrivalRecordedInCas1
---> PlacementApplication(254), withdrawable:Y, mayDirectlyWithdraw:Y
------> PlacementRequest(bbf), withdrawable:Y, mayDirectlyWithdraw:N
---------> Booking(76b), withdrawable:Y, mayDirectlyWithdraw:Y
------> PlacementRequest(0b3), withdrawable:Y, mayDirectlyWithdraw:N
---> PlacementApplication(835), withdrawable:Y, mayDirectlyWithdraw:Y
---> Booking(bde), withdrawable:Y, mayDirectlyWithdraw:Y

Notes: [1 or more placements cannot be withdrawn as they have an arrival]