Check arrival status in delius on CAS1 Withdrawal
When building the Cas1 Withdrawable Tree for a placement we now check the status of the placement in delius wrt. arrivals. If the placement has been marked as arrived in delius we ‘block’ the withdrawal as we would if the placement was marked as arrived in CAS1. We also include a specific note in the withdrawal tree indicating that 1 or placements have arrivals marked in delius and as such can’t be withdrawn.

Use of this functionality is controlled by the “cas1-check-delius-withdrawal-status” feature flag, which will return false by default

There is no caching on the delius referrals endpoint that we use to retrieve this data. Caching wouldn’t make sense in this context because we always want the latest state wrt. arrivals. If we find that the endpoint is being invoked too often we could look at either

1) caching a response that indicates there is an arrival (assuming this is easy with redis integration)
2) write the delius arrival on the booking entity itself (or via a new table)

Both of these approaches assume that once marked as a arrived in delius this can’t be undone (e.g. if the person was marked as arrived by accident)